+++ src/lib/lms-index.ts
@@ -2,9 +2,6 @@
 import fs from 'fs';
 import matter from 'gray-matter';
 import path from 'path';
-import lms from 'src/lib/index.json';
-
-import { filterIndex } from './lms-search';
 
 
 export interface indexField {
@@ -15,18 +12,6 @@ export interface indexField {
     metaDescription?: string,
     access: string,
     tags?: string[]
-}
-
-export interface filteredIndex {
-    field: {
-        path: string,
-        navTitle?: string,
-        title: string,
-        metaTitle: string,
-        metaDescription?: string,
-        access: string,
-    },
-    tags: string[]
 }
 
 export async function createIndexArray(dir: string) {
@@ -62,15 +47,3 @@ export async function createIndexArray(dir: string) {
 
     fs.writeFileSync('public/lms/assets/lmsindex.json', JSON.stringify(index));
 };
-
-export async function searchInput(input: string) {
-  const input2 = input.split(',');
-    const filteredIndex: Array<filteredIndex> = [];
-    for (let i = 0; i < lms.length; i++){
-      const inputs = filterIndex(lms[i], input2);
-      if (inputs) {
-        filteredIndex.push(inputs);
-      };
-    };
-    return filteredIndex
-}
